backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
groq>=0.9.0
edge-tts>=6.1.15
aiofiles==23.2.1
python-dotenv==1.0.1
httpx>=0.26.0,<0.29.0
supabase>=2.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator>=2.0.0
